create testing module, add to docs
