+++ docs/api/testing/cases.md
@@ -1,0 +1,7 @@
+# cases
+
+```{eval-rst}
+.. automodule:: numpydantic.testing.cases
+    :members:
+    :undoc-members:
+```
