+++ docs/api/testing/helpers.md
@@ -1,0 +1,7 @@
+# helpers
+
+```{eval-rst}
+.. automodule:: numpydantic.testing.helpers
+    :members:
+    :undoc-members:
+```
