+++ docs/api/testing/index.md
@@ -1,0 +1,8 @@
+# testing
+
+Utilities for testing and 3rd-party interface development.
+
+```{toctree}
+cases
+helpers
+```
